First parse secrets from environment variables
Secrets are first parsed from environment variables and if they're not
present and values in the 'tara.toml' are used. If the tara.toml isn't
found the secrets are attempted to be parsed from the environment.
Environment variables are parsed from any .env files in the CWD or any
parent directories via dotenvy.